Add custom domain name
Deploying to production erected a couple of obstacles not found in dev.
For one, the resource bindings we're creating in time for the next build
step. We got around this by deploying to dev first, then reusuing the
same environment for the prod deployment. There's an active issue
regarding this, but in the meantime it might be best to stop using
resource bindings.

The second issue was strange TypeError received when using our
next.config permanent login redirect. Still haven't figured out why this
works in dev and not production but we've gotten around it by
redirecting to the Cognito login URL directly from the client side.